feat(standard response): Created standard response, implemented in default server module
